requirements.txt
discord.py==1.7.1
requests==2.23.0
pymongo==3.11.3